# Back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation",
        "contact": {
            "name": "Soporte",
            "email": "soporte@terciariourquiza.edu.ar"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/aulas": {
            "get": {
                "tags": [
                    "Aula"
                ],
                "summary": "Obtener todas las aulas",
                "description": "Devuelve todas las aulas",
                "operationId": "getAulas",
                "responses": {
                    "200": {
                        "description": "Aulas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Aula"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener las aulas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Aula"
                ],
                "summary": "Guardar un aula",
                "description": "Guardar un aula",
                "operationId": "guardarAula",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Aula"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aula guardada correctamente"
                    },
                    "400": {
                        "description": "Error al guardar el aula"
                    }
                }
            }
        },
        "/api/aulas/{id}": {
            "get": {
                "tags": [
                    "Aula"
                ],
                "summary": "Obtener un aula por id",
                "description": "Obtener un aula por id",
                "operationId": "obtenerAulaPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del aula",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aula obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Aula"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontró el aula"
                    },
                    "500": {
                        "description": "Error al obtener el aula"
                    }
                }
            }
        },
        "/api/aulas/actualizar/{id}": {
            "put": {
                "tags": [
                    "Aula"
                ],
                "summary": "Actualizar un aula",
                "description": "Actualizar un aula",
                "operationId": "actualizarAula",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del aula",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Aula"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aula actualizada correctamente"
                    },
                    "500": {
                        "description": "Error al actualizar el aula"
                    }
                }
            }
        },
        "/api/aulas/eliminar/{id}": {
            "delete": {
                "tags": [
                    "Aula"
                ],
                "summary": "Eliminar un aula",
                "description": "Eliminar un aula",
                "operationId": "eliminarAula",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del aula",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aula eliminada correctamente"
                    },
                    "500": {
                        "description": "Error al eliminar el aula"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Aula": {
                "title": "Aula",
                "description": "Esquema del objeto Aula",
                "properties": {
                    "id_aula": {
                        "description": "ID del aula",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre del aula",
                        "type": "string"
                    },
                    "capacidad": {
                        "description": "Capacidad del aula",
                        "type": "integer"
                    },
                    "tipo_aula": {
                        "description": "Tipo de aula",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}